changed trajectories_api structure, moved linear to it's own module
