+++ blender3d_intergration/interfaces.py
@@ -3,8 +3,7 @@ from abc import abstractmethod, ABC
 from matplotlib import pyplot as plt
 
 from blender3d_intergration.trajectories_api.models import Trajectory, \
-    Coordinates, CircularTrajectoryMotionParameters, CircularTrajectoryMotionInitialParameters, FramesPerSecond, \
-    LinearTrajectoryMotionInitialParameters, LinearTrajectoryMotionParameters
+    Coordinates, CircularTrajectoryMotionParameters, CircularTrajectoryMotionInitialParameters, FramesPerSecond
 
 
 class TrajectoryCalculator:
@@ -66,21 +65,3 @@ class CircleTrajectory(TrajectoryCalculator, ABC):
     def get_motion_parameters(self) -> CircularTrajectoryMotionParameters:
         return self._motion_parameters
 
-
-class LinearTrajectory(TrajectoryCalculator, ABC):
-    _initial_motion_parameters: LinearTrajectoryMotionInitialParameters
-    _motion_parameters: LinearTrajectoryMotionParameters
-
-    def __init__(self,
-                 initial_motion_parameters: LinearTrajectoryMotionInitialParameters,
-                 fps: FramesPerSecond):
-        self._initial_motion_parameters = initial_motion_parameters
-        self._motion_parameters = LinearTrajectoryMotionParameters(self._initial_motion_parameters)
-        self._fps = fps
-        self._trajectory = Trajectory()
-
-    def get_initial_motion_parameters(self) -> LinearTrajectoryMotionInitialParameters:
-        return self._initial_motion_parameters
-
-    def get_motion_parameters(self) -> LinearTrajectoryMotionParameters:
-        return self._motion_parameters
