+++ blender3d_intergration/trajectories_generation.py
@@ -1,11 +1,7 @@
 import os
 
 from blender3d_intergration.blender_python_working.blender_commands_generator import TrajectoryToBlenderCommands
-from blender3d_intergration.trajectories_api.linear_trajectory import SimpleLinearTrajectory
-from blender3d_intergration.trajectories_api.models import CircularTrajectoryMotionInitialParameters, \
-    FramesPerSecond, Coordinates, LinearTrajectoryMotionInitialParameters
-from blender3d_intergration.trajectories_api.trajectory_calculation_automations import \
-    circular_trajectories_constant_height
+from blender3d_intergration.trajectories_api.models import FramesPerSecond, Coordinates
 
 radius_values_list = [20, 30, 40, 50, 60, 70, 80, 90, 100]
 height_values_list = [30, 40, 50, 60, 70, 80, 90, 100]
@@ -13,7 +9,7 @@ height_values_list = [30, 40, 50, 60, 70, 80, 90, 100]
 linear_velocity_m_s = 18
 
 trajectories_path = 'trajectories_api/calculated_trajectories/new'
-linear_trajectories_path = 'trajectories_api/calculated_trajectories/linear'
+linear_trajectories_path = 'trajectories_api/calculated_trajectories/linear/test'
 script_dir = os.path.dirname(__file__)
 full_path = os.path.join(script_dir, trajectories_path)
 linear_full_path = os.path.join(script_dir, linear_trajectories_path)
@@ -42,24 +38,3 @@ linear_full_path = os.path.join(script_dir, linear_trajectories_path)
 #             path_to_files=full_path,
 #             output_file_name=f"h{height}m_r{radius}m",
 #         )
-
-for height in height_values_list:
-    start_point = Coordinates(x=10, y=0, z=height, frame=0, time_s=0),
-    end_point = Coordinates(x=100, y=0, z=height, frame=0, time_s=0),
-
-    linear_initial_params = LinearTrajectoryMotionInitialParameters(
-        start_point=Coordinates(x=10, y=0, z=30, frame=0, time_s=0),
-        end_point=Coordinates(x=100, y=0, z=30, frame=0, time_s=0),
-        x_velocity_m_s=18,
-        y_velocity_m_s=0,
-        z_velocity_m_s=0
-    )
-
-    linear_trajectory = SimpleLinearTrajectory(linear_initial_params, FramesPerSecond.FPS_30)
-    linear_trajectory.calculate_trajectory()
-
-    command_printer = TrajectoryToBlenderCommands.to_bpy(
-        trajectory=linear_trajectory.get_trajectory(),
-        path_to_files=linear_full_path,
-        output_file_name=f"linear_z{height}m",
-    )
